Use insertLookupWithKey when needing old value.
updateLookupWithKey returns original value only if it was deleted.  If
it was updated, it returns the updated value.  On both acid-state and
MockStorage we want the old value, though.

Now all acid-state tests are green once more.

MockStorage tests were already green because MockStorage does not
mantain a cached mapping of auth IDs to session IDs.