Travis: Use PostgreSQL 9.3.
The persistent test is failing and I have no idea why, as it's not
displaying any error message.  This commit changes a few things with
PostgreSQL just to see if they make any difference.

Note that the persistent test suite passes with flying colors both
before and after this commit on my machine using PostgreSQL 9.4.2.