Update our build matrix
Looks like we hit a big breaking change with the latest yesod major
version. I took a look at the CPP needed and it wouldn't be worth
it. This forces us to cut lts-10 and older loose. I've also allowed
nightly tests to fail since there are a lot of dependencies that are
trailing behind.